add minus point logic

+++ MyPolyglotWeb/Models/DomainModels/ScoreDB.cs
@@ -4,6 +4,6 @@
     {
         public virtual LessonDB Lesson { get; set; }
         public virtual UserDB User { get; set; }
-        public virtual int Points { get; set; }
+        public virtual double Points { get; set; }
     }
 }
